# protocol/validation: optimize CalcMerkleRoot